Make 'stretch' aligned children fill cross-axis in containers of undefined size in the cross-axis
Summary: Fetched and rebased
https://github.com/facebook/css-layout/pull/145 and updated it to do 2
things:
1) Make 'stretch' aligned children grow to fill their containers in the
cross-axis if the cross-axis size is undefined. The added test might be
the best explanation of this.
2) Make sure this doesn’t decrease the frequency of the
“simpleStackCross” optimization in a relatively complex sample
environment.

Test Plan: Added a test: Container with 3 row directioned children and
no explicit height set. One child is of height 0, one of 0 height with
stretch alignment, and one of 150 height. Expected the 0 height child
remains of height 0, stretch grows to 150, and 150 stays the same.